% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/houseman.R
\name{estimate.cell.proportions}
\alias{estimate.cell.proportions}
\title{Estimate cell-type coefficients. Returns matrix of cell-type proportions.}
\usage{
estimate.cell.proportions(B, top_n = 500, mc.cores = 2, cell.coefs = NULL)
}
\arguments{
\item{B}{matrix of beta values (n_probes * n_samples) rownames must
be the illumina cg id or the chrom:position}

\item{top_n}{number of probes from cell.coefs to use}

\item{mc.cores}{number of cores to use for parallelization}

\item{cell.coefs}{path to tab-delimited file containind DMR-probes for the
cell types of interest. system.file("extdata", "houseman-dmrs.txt",
package="celltypes450")}
}
\value{
matrix of estimated cell-type proportions
}
\description{
Estimate cell-type coefficients. Returns matrix of cell-type proportions.
}

