% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/houseman.R
\name{m.ilogit}
\alias{m.ilogit}
\title{Perform the anti-logit}
\usage{
m.ilogit(M)
}
\arguments{
\item{M}{matrix of M values on which to apply anti-logit.}
}
\value{
beta matrix
}
\description{
Perform the anti-logit
}

