% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/houseman.R
\name{m.logit}
\alias{m.logit}
\title{Perform the logit transform.}
\usage{
m.logit(b)
}
\arguments{
\item{b}{matrix of beta values on which to apply logit.}
}
\value{
M matrix
}
\description{
Perform the logit transform.
}

