% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/houseman.R
\name{adjust.beta}
\alias{adjust.beta}
\title{Adjust Beta for cell mixture. Returns beta of the average cell-type.}
\usage{
adjust.beta(B, top_n = 500, mc.cores = 2, cell.coefs = NULL,
  omega.mix = NULL)
}
\arguments{
\item{B}{matrix of beta values (n_probes * n_samples) rownames must
be the illumina cg id or the chrom:position}

\item{top_n}{number of probes from cell.coefs to use}

\item{mc.cores}{number of cores to use for parallelization}

\item{cell.coefs}{path to tab-delimited file containind DMR-probes for the
cell types of interest. Ignored if omega.mix is provided.
system.file("extdata", "houseman-dmrs.txt", package="celltypes450")}

\item{omega.mix}{pre-specified matrix of cell-type coefficients. If
not specified, this will be estimated from the data.}
}
\value{
matrix of adjusted beta values
}
\description{
For simply estimating the cell-type proportions, see
estimate.cell.proportions.
}

